added CardWrapper and CardComponent into separate files, reusable components to make the code clean
